+++ src/main/java/org/delivery/KiwiEats/controllers/ProductController.java
@@ -2,7 +2,7 @@ package org.delivery.KiwiEats.controllers;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.delivery.KiwiEats.exceptions.NotFoundException;
+import org.delivery.KiwiEats.exception.NotFoundException;
 import org.delivery.KiwiEats.models.ProductDTO;
 import org.delivery.KiwiEats.services.ProductService;
 import org.springframework.http.ResponseEntity;
