+++ src/main/java/org/delivery/KiwiEats/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package org.delivery.KiwiEats.exceptions;
+package org.delivery.KiwiEats.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
